+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- The core Event class accepts `causation_id` to allow event stores to
+  add support for tracking causation ids with events.
+- The core Memory event store saves the `causation_id`.
+- The event store shared RSpec examples specify event stores should save
+  the `causation_id`.
+
 ### Removed
 - The `processing_event` method from the memory tracker. It was intended to
   be a mechanism to wrap processing and tracker updates which appears to be